Reorganized tests into generic file row scope tests and trait importer validator tests
